Change type of version in person, role and privilege
For this also changed the type of version to long in their respective HBM.

+++ smart-user-api/src/main/java/com/smartitengineering/user/domain/Privilege.java
@@ -13,7 +13,7 @@ import org.codehaus.jackson.annotate.JsonIgnore;
  *
  * @author modhu7
  */
-public class Privilege extends AbstractGenericPersistentDTO<Privilege, Long, Integer> {
+public class Privilege extends AbstractGenericPersistentDTO<Privilege, Long, Long> {
 
   private Organization parentOrganization;
   private String name;
